Separate ExtensionManager from QuPathGUI
